Refactored GPU/Keypoint detector/Descriptor extractor
git-svn-id: http://find-object.googlecode.com/svn/trunk/find_object@342 620bd6b2-0a58-f614-fd9a-1bd335dccda9

+++ imagesTcpServer/ImagesTcpServer.cpp
@@ -66,7 +66,7 @@ void ImagesTcpServer::publishImage(const cv::Mat & image)
 			QDataStream out(&block, QIODevice::WriteOnly);
 			out.setVersion(QDataStream::Qt_4_0);
 			out << (quint64)0;
-			out.writeRawData((char*)buf.data(), buf.size());
+			out.writeRawData((char*)buf.data(), (int)buf.size());
 			out.device()->seek(0);
 			out << (quint64)(block.size() - sizeof(quint64));
 			(*iter)->write(block);
